requirements.txt
pillow
matplotlib
wget
numpy==1.16.0